apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai==1.107.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-dotenv==1.1.1
pypdf==5.9.0
Pillow==11.3.0
python-dateutil==2.9.0.post0
